# openapi.json
{"openapi": "3.1.0", "info": {"title": "RPG Hell API", "description": "API for managing all of the data that exists in the RPG Hell tabletop game", "version": "0.1.0"}, "paths": {"/login": {"get": {"tags": ["Users"], "summary": "Login", "operationId": "login", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/logout": {"get": {"tags": ["Users"], "summary": "Logout", "operationId": "logout", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/auth/discord/callback": {"get": {"tags": ["Users"], "summary": "Callback", "operationId": "callback", "parameters": [{"name": "code", "in": "query", "required": true, "schema": {"type": "string", "title": "Code"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/refresh": {"get": {"tags": ["Users"], "summary": "Refresh", "operationId": "refresh", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/user": {"get": {"tags": ["Users"], "summary": "Get Discord User", "operationId": "get_discord_user", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/me": {"get": {"tags": ["Users"], "summary": "Get Or Create Database User", "operationId": "get_or_create_database_user", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DBUser"}}}}}}}, "/trait/": {"put": {"tags": ["Traits"], "summary": "Put Trait", "operationId": "put_trait", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Trait"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/traits/": {"get": {"tags": ["Traits"], "summary": "Get All Traits", "operationId": "get_all_traits", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"$ref": "#/components/schemas/Trait"}, "type": "object", "title": "Response Traits-Get All Traits"}}}}}}}, "/trait/{name}": {"patch": {"tags": ["Traits"], "summary": "Update Trait", "operationId": "update_trait", "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string", "title": "Name"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Trait"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/trait/{id}": {"delete": {"tags": ["Traits"], "summary": "Delete Trait", "operationId": "delete_trait", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/filtertraits/": {"get": {"tags": ["Traits"], "summary": "Filter Traits By Requirements", "operationId": "filter_traits_by_requirements", "parameters": [{"name": "requirements", "in": "query", "required": true, "schema": {"type": "string", "description": "csv", "title": "Requirements"}, "description": "csv"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": {"$ref": "#/components/schemas/Trait"}, "title": "Response Traits-Filter Traits By Requirements"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/item/": {"put": {"tags": ["Items"], "summary": "Put Item", "operationId": "put_item", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Item"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/items/": {"get": {"tags": ["Items"], "summary": "Get All Items", "operationId": "get_all_items", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"$ref": "#/components/schemas/Item"}, "type": "object", "title": "Response Items-Get All Items"}}}}}}}, "/item/{name}": {"patch": {"tags": ["Items"], "summary": "Update Item", "operationId": "update_item", "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string", "title": "Name"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Item"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/item/{id}": {"delete": {"tags": ["Items"], "summary": "Delete Item", "operationId": "delete_item", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/filteritems/{option}": {"get": {"tags": ["Items"], "summary": "Filter Items By Tags Or Requirements", "operationId": "filter_items_by_tags_or_requirements", "parameters": [{"name": "option", "in": "path", "required": true, "schema": {"allOf": [{"$ref": "#/components/schemas/FilterOption"}], "description": " 0 = tags, 1 = requirements", "title": "Option"}, "description": " 0 = tags, 1 = requirements"}, {"name": "filters", "in": "query", "required": true, "schema": {"type": "string", "description": "csv", "title": "Filters"}, "description": "csv"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": {"$ref": "#/components/schemas/Item"}, "title": "Response Items-Filter Items By Tags Or Requirements"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/spell/": {"get": {"tags": ["Spells"], "summary": "Spell Search", "description": "This method will eventually support filtering all values. Not just name.", "operationId": "spell_search", "parameters": [{"name": "name", "in": "query", "required": true, "schema": {"type": "string", "title": "Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Spell"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Spells", "Spells"], "summary": "Put Spell", "operationId": "put_spell", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Spell"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/spells/": {"get": {"tags": ["Spells"], "summary": "Get All Spells", "operationId": "get_all_spells", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"$ref": "#/components/schemas/Spell"}, "type": "object", "title": "Response Spells-Get All Spells"}}}}}}}, "/spell/{id}": {"get": {"tags": ["Spells"], "summary": "Get Spell By Id", "operationId": "get_spell_by_id", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "The ID of the spell to get"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Spell"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Spells"], "summary": "Delete Spell", "operationId": "delete_spell", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/spell/{name}": {"patch": {"tags": ["Spells"], "summary": "Update Spell", "operationId": "update_spell", "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string", "title": "Name"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Spell"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/filterspells/": {"get": {"tags": ["Spells"], "summary": "Filter Spells By Tags", "operationId": "filter_spells_by_tags", "parameters": [{"name": "tags", "in": "query", "required": true, "schema": {"type": "string", "description": "csv", "title": "Tags"}, "description": "csv"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": {"$ref": "#/components/schemas/Spell"}, "title": "Response Spells-Filter Spells By Tags"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/object/": {"get": {"tags": ["Objects"], "summary": "Object Search", "operationId": "object_search", "parameters": [{"name": "name", "in": "query", "required": true, "schema": {"type": "string", "title": "Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/Trait"}, {"$ref": "#/components/schemas/Item"}], "title": "Response Objects-Object Search"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/object/{id}": {"get": {"tags": ["Objects"], "summary": "Get Object By Id", "operationId": "get_object_by_id", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "The ID of the object to get"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/Trait"}, {"$ref": "#/components/schemas/Item"}], "title": "Response Objects-Get Object By Id"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/creature/": {"get": {"tags": ["Creatures"], "summary": "Creature Search", "description": "This method will eventually support filtering all values. Not just name.", "operationId": "creature_search", "parameters": [{"name": "name", "in": "query", "required": true, "schema": {"type": "string", "title": "Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Creature"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Creatures"], "summary": "Put Creature", "operationId": "put_creature", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Creature"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/creature/{id}": {"get": {"tags": ["Creatures"], "summary": "Get Creature By Id", "operationId": "get_creature_by_id", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "The ID of the creature to get"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Creature"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"tags": ["Creatures"], "summary": "Update Creature", "operationId": "update_creature", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Creature"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Creatures"], "summary": "Delete Creature", "operationId": "delete_creature", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"tags": ["Temporary"], "summary": "Root", "operationId": "root", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/openquery": {"put": {"tags": ["Temporary"], "summary": "Open Query", "description": "Takes in a raw SQL query.", "operationId": "open_query", "parameters": [{"name": "query", "in": "query", "required": true, "schema": {"type": "string", "title": "Query"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Creature": {"properties": {"id": {"type": "integer", "title": "Id", "default": 0}, "name": {"type": "string", "title": "Name"}, "types": {"items": {"type": "string"}, "type": "array", "title": "Types", "default": []}, "level": {"type": "integer", "title": "Level", "default": 0}, "body": {"type": "integer", "title": "Body", "default": 0}, "mind": {"type": "integer", "title": "Mind", "default": 0}, "soul": {"type": "integer", "title": "Soul", "default": 0}, "arcana": {"type": "integer", "title": "Arcana", "default": 0}, "crafting": {"type": "integer", "title": "Crafting", "default": 0}, "charm": {"type": "integer", "title": "Charm", "default": 0}, "thieving": {"type": "integer", "title": "Thieving", "default": 0}, "nature": {"type": "integer", "title": "Nature", "default": 0}, "medicine": {"type": "integer", "title": "Medicine", "default": 0}, "traits": {"items": {"type": "string"}, "type": "array", "title": "Traits", "default": []}, "spells": {"items": {"type": "string"}, "type": "array", "title": "Spells", "default": []}, "items": {"items": {"type": "string"}, "type": "array", "title": "Items", "default": []}, "notes": {"type": "string", "title": "Notes", "default": ""}}, "type": "object", "required": ["name"], "title": "Creature"}, "DBUser": {"properties": {"id": {"type": "integer", "title": "Id", "default": 0}, "discord_id": {"type": "integer", "title": "Discord Id"}, "username": {"type": "string", "title": "Username"}, "email": {"type": "string", "title": "Email"}, "is_admin": {"type": "boolean", "title": "Is Admin", "default": false}, "avatar_url": {"type": "string", "title": "Avatar Url", "default": ""}}, "type": "object", "required": ["discord_id", "username", "email"], "title": "DBUser"}, "FilterOption": {"type": "string", "enum": ["0", "1"], "title": "FilterOption"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Item": {"properties": {"id": {"type": "integer", "title": "Id", "default": 0}, "name": {"type": "string", "title": "Name"}, "effect": {"type": "string", "title": "Effect"}, "req": {"items": {"type": "string"}, "type": "array", "title": "Req", "default": []}, "cost": {"type": "integer", "title": "Cost", "default": 0}, "craft": {"type": "integer", "title": "Craft", "default": 0}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags", "default": []}}, "type": "object", "required": ["name"], "title": "Item"}, "Spell": {"properties": {"id": {"type": "integer", "title": "Id", "default": 0}, "name": {"type": "string", "title": "Name"}, "effect": {"type": "string", "title": "Effect"}, "dice": {"type": "integer", "title": "Dice", "default": 0}, "level": {"type": "integer", "title": "Level", "default": 0}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags", "default": []}}, "type": "object", "required": ["name"], "title": "Spell"}, "Trait": {"properties": {"id": {"type": "integer", "title": "Id", "default": 0}, "name": {"type": "string", "title": "Name"}, "effect": {"type": "string", "title": "Effect"}, "req": {"items": {"type": "string"}, "type": "array", "title": "Req", "default": []}, "dice": {"type": "integer", "title": "Dice"}, "is_passive": {"type": "boolean", "title": "Is Passive", "default": false}}, "type": "object", "required": ["name", "dice"], "title": "Trait"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}, "tags": [{"name": "Temporary", "description": "A temporary endpoint that will not exist when this api has been depoloyed to production."}, {"name": "Objects", "description": ""}, {"name": "Traits", "description": ""}, {"name": "Spells", "description": ""}, {"name": "Items", "description": ""}, {"name": "Creatures", "description": ""}, {"name": "Users / Authentication", "description": ""}]}